+++ src/locales/default/metadata.ts
@@ -3,10 +3,30 @@ export default {
     description: '{{appName}} 带给你最好的 ChatGPT, Claude , Gemini, OLLaMA WebUI 使用体验',
     title: '{{appName}}：个人 AI 效能工具，给自己一个更聪明的大脑',
   },
-  market: {
+  discover: {
+    assistants: {
+      description:
+        '内容创作、文案、问答、图像生成、视频生成、语音生成、智能 Agent、自动化工作流，定制你专属的 AI / GPTs / OLLaMA 智能助手',
+      title: 'AI助手',
+    },
     description:
-      '内容创作、文案、问答、图像生成、视频生成、语音生成、智能 Agent、自动化工作流，定制你专属的 AI 智能助手',
-    title: '助手市场',
+      '内容创作、文案、问答、图像生成、视频生成、语音生成、智能 Agent、自动化工作流、自定义AI应用，定制你专属的 AI 应用工作台',
+    models: {
+      description: '探索主流 AI 模型 OpenAI / GPT / Claude 3 / Gemini / Ollama / Azure / DeepSeek',
+      title: 'AI模型',
+    },
+    plugins: {
+      description:
+        '搜素图表生成、学术、图像生成、视频生成、语音生成、自动化工作流，为你的助手集成丰富的插件能力',
+      title: 'AI插件',
+    },
+    providers: {
+      description:
+        '探索主流模型供应商 OpenAI / Qwen / Ollama / Anthropic / DeepSeek / Google Gemini / OpenRouter',
+      title: 'AI模型服务商',
+    },
+    search: '搜索',
+    title: '发现',
   },
   plugins: {
     description: